adding components to the user page
