+++ client/src/components/userProfile/favoriteBreeds.jsx
@@ -6,7 +6,9 @@ const FavoriteDogs = () => {
   return (
     <div>
       {/* map the dataset of dogs */}
-      <DogCard orientation='portrait' />
+      <ul>
+        <DogCard orientation='portrait' />
+      </ul>
     </div>
   )
 }
