+++ client/src/components/userProfile/userMenu.jsx
@@ -1,9 +1,19 @@
-import React from 'react';
+import React, { useState } from 'react';
+
 import Button from '@mui/material/Button';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 
 const UserMenu = () => {
+  const [anchorEl, setAnchorEl] = useState(null);
+  const open = Boolean(anchorEl);
+  const handleClick = (e) => {
+    setAnchorEl(e.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  }
+
   return (
     <div>
       <Button
@@ -29,3 +39,5 @@ const UserMenu = () => {
     </div>
   )
 }
+
+export default UserMenu;
