+++ client/src/components/userProfile/userPage.jsx
@@ -3,26 +3,26 @@ import FavoriteDogs from './favoriteDogs.jsx';
 import FavoriteBreeds from './favoriteBreeds.jsx';
 import SideBar from './sideBar.jsx';
 
+import Grid from '@mui/material/Grid';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
 
 const UserPage = () => {
-  const [anchorEl, setAnchorEl] = useState(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (e) => {
-    setAnchorEl(e.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  }
-
-
-
   return (
     <Container className='innerContainer'>
-      <SideBar />
-      <FavoriteBreeds />
-      <FavoriteDogs />
+      <Grid container spacing={2}>
+        <Grid item>
+          <SideBar />
+        </Grid>
+        <Grid container>
+          <Grid item>
+            <FavoriteBreeds />
+          </Grid>
+          <Grid item>
+            <FavoriteDogs />
+          </Grid>
+        </Grid>
+      </Grid>
     </Container>
   )
 }
